refactor: created a shared folder for reusable panel components
